Module Cobol_indent_old.Indent_util

val check_pos : Cobol_preproc.Src_format.any -> Cobol_common.Srcloc.srcloc -> int -> Indent_type.indent_record list -> bool -> Indent_type.indent_record list

val failure_msg : Cobol_common.Srcloc.srcloc -> string

val offset_of_keyword : Indent_type.indent_config -> Indent_type.context_kind -> int

val offset_of_context : Indent_type.context -> int

val push_context : Indent_type.indent_state -> Indent_type.context_kind -> Indent_type.context -> Indent_type.context

val is_data_decl : string -> bool

val reduce_level : int -> Indent_type.context -> Indent_type.context

val handle_period : Indent_type.context -> Indent_type.context

val exp_scope_termination : Indent_type.context_kind -> Indent_type.context -> Indent_type.context

val pop_until : Indent_type.context_kind -> Indent_type.context -> Indent_type.context

val pop_until_division : Indent_type.context -> Indent_type.context

val pop_until_compilation_unit : Indent_type.context -> Indent_type.context

val imp_scope_termination : Indent_type.context -> Indent_type.context

val phrase_termination : Indent_type.context -> Indent_type.context

val phrase_termination_until : Indent_type.context_kind -> Indent_type.context -> Indent_type.context

val apply : string -> Indent_type.indent_record list -> string

apply s rdl applies the changes in rdl to the text in s. rdl is expected to have been computed by calling indent_range ~contents:s.

Note: this function is currently untested. It is only used by the command `superbol indent`. The main way to interact with the indenter is through LSP requests, which are tested in lsp_formatting.ml